refactor: increase testability
