+++ src/ec2/commands.ts
@@ -3,38 +3,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { createQuickPick, QuickPickPrompter } from "../shared/ui/pickerPrompter"
-import { DataQuickPickItem } from "../shared/ui/pickerPrompter"
-import { AsyncCollection } from "../shared/utilities/asyncCollection"
-import { isValidResponse } from '../shared/wizards/wizard'
 import { EC2 } from 'aws-sdk'
+import globals from '../shared/extensionGlobals'
+import { pageableToCollection } from '../shared/utilities/collectionUtils'
+import { CancellationError } from '../shared/utilities/timeoutUtils'
+import { extractInstanceIdsFromReservations } from "./utils"
+import { selectInstance } from './prompter'
 
-function asQuickpickItem(stringItem: string): DataQuickPickItem<string>[] 
-{
-    return [{
-        label: stringItem, 
-        data: stringItem
-    }]
+export async function attemptConnection(defaultRegion: string): Promise<void> {
+    const client = await globals.sdkClientBuilder.createAwsService(EC2, undefined, defaultRegion)
+            const requester = async (request: EC2.DescribeInstancesRequest) => 
+                client.describeInstances(request).promise() 
+            const collection = extractInstanceIdsFromReservations(pageableToCollection(requester, {}, 'NextToken', 'Reservations'))
+
+            const selection = await selectInstance(collection)
+            if(!selection){
+                throw new CancellationError('user')
+            } else {
+                console.log(selection)
+            }
 }
-
-function createInstancePrompter(instances: AsyncCollection<string>): QuickPickPrompter<string> {
-    const items = instances.map(asQuickpickItem)
-    const prompter = createQuickPick(items, {title: "Select EC2 instance by id"})
-
-    return prompter
-}
-
-export async function selectInstance(instances: AsyncCollection<string>): Promise<string | undefined> {
-    const prompter = createInstancePrompter(instances)
-    const response = await prompter.prompt() 
-    return isValidResponse(response) ? response : undefined
-}
-
-export function extractInstanceIds(reservations: AsyncCollection<EC2.ReservationList | undefined>): AsyncCollection<string> {
-    return reservations
-        .flatten()
-        .map(instanceList => instanceList?.Instances)
-        .flatten()
-        .map(instance => instance?.InstanceId)
-        .filter(instanceId => instanceId !== undefined)
-}
